+++ example/lib/model.dart
@@ -1,5 +1,0 @@
-class CounterModel {
-  final int count;
-
-  CounterModel({this.count = 0});
-}